refactor(e2e): Split 'pkg' tests to separate files
